+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.6] - 2024-08-29
+
+### Changed
+
+- Changed the **Indirect Diffuse Lighting Multiplier** property type to always visible.
+
+### Fixed
+
+- Fixed an issue where deferred shader variants were stripped by URP 14 during the build.
+- Fixed an issue with SSGI on mobile devices.
+
+
 ## [1.0.5] - 2024-08-01
 
 ### Changed